Conflicts:
	Scribus/scribus/plugins/gettext/sxwim/contentreader.cpp
	Scribus/scribus/plugins/import/cgm/importcgmplugin.cpp
	Scribus/scribus/scribusview.cpp
	Scribus/scribus/text/storytext.cpp

+++ Scribus/scribus/plugins/gettext/sxwim/contentreader.cpp
@@ -126,8 +126,8 @@ bool ContentReader::startElement(const QString&, const QString&, const QString &
 	else if (name == "text:list-item")
 	{
 		bool isOrdered = false;
-		long levelIndex = listLevel - 1;
-		if (levelIndex >= 0 && levelIndex < (long) isOrdered2.size())
+		int levelIndex = listLevel - 1;
+		if (levelIndex >= 0 && levelIndex < static_cast<int>(isOrdered2.size()))
 			isOrdered = isOrdered2[listLevel - 1];
 		if (isOrdered)
 		{
